# packages/server/api-spec.yaml
openapi: 3.0.3
info:
  title: ZypherAgent API
  description: |
    API for controlling the ZypherAgent to perform code editing tasks.
    This API allows frontend applications to interact with the ZypherAgent,
    send tasks, and receive responses.

    The API server creates a single ZypherAgent instance at startup with
    pre-specified working directory and configuration.
  version: 1.0.0
  contact:
    name: ZypherAgent Team
servers:
  - url: "http://localhost:3000"
    description: Local development server

components:
  schemas:
    Error:
      type: object
      required:
        - code
        - type
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        type:
          type: string
          description: Error type (e.g. invalid_request, internal_server_error)
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
        error:
          type: string
          description: Formatted error message for internal server errors

    ContentBlock:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [text, tool_use, tool_result]
          description: The type of content block
        text:
          type: string
          description: Text content (when type is 'text')
        name:
          type: string
          description: Tool name (when type is 'tool_use')
        input:
          type: object
          description: Tool parameters (when type is 'tool_use')
        tool_use_id:
          type: string
          description: ID of the tool use (when type is 'tool_result')
        content:
          type: string
          description: Tool result content (when type is 'tool_result')

    Checkpoint:
      type: object
      required:
        - id
        - name
        - timestamp
      properties:
        id:
          type: string
          description: Unique identifier for the checkpoint (Git commit hash)
        name:
          type: string
          description: User-friendly name for the checkpoint
        timestamp:
          type: string
          format: date-time
          description: When the checkpoint was created
        files:
          type: array
          items:
            type: string
          description: Files changed in this checkpoint

    Message:
      type: object
      required:
        - role
        - content
        - timestamp
      properties:
        role:
          type: string
          enum: [user, assistant]
          description: The role of the message sender
        content:
          oneOf:
            - type: string
            - type: array
              items:
                $ref: "#/components/schemas/ContentBlock"
          description: The content of the message
        timestamp:
          type: string
          format: date-time
          description: When the message was created
        checkpointId:
          type: string
          description: Reference to a checkpoint created before this message (for user task messages)
        checkpoint:
          $ref: "#/components/schemas/Checkpoint"
          description: Metadata about the checkpoint

    TaskRequest:
      type: object
      required:
        - task
      properties:
        task:
          type: string
          description: The task description to send to the agent
        imageAttachments:
          type: array
          items:
            type: object
            required:
              - name
              - data
            properties:
              name:
                type: string
                description: Name of the image file
              data:
                type: string
                description: Image data in format data:${mimeType};base64,${base64String}
                example: "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEASABIAAD/..."
          description: Image attachments for the task

    StreamEvent:
      type: object
      required:
        - event
        - data
      properties:
        event:
          type: string
          enum: [
            message,
            content_delta,
            tool_use_delta,
            complete,
            error,
            heartbeat,
            cancelled,
          ]
          description: The type of event
        data:
          oneOf:
            - $ref: "#/components/schemas/Message"
            - type: object
              properties:
                content:
                  type: string
                  description: Content delta for incremental updates
            - type: object
              properties:
                name:
                  type: string
                  description: Tool name
                partialInput:
                  type: object
                  description: Partial tool input parameters
            - type: object
              description: Empty object for complete event
            - type: object
              properties:
                error:
                  type: string
                  description: Error message
              description: Error event data
            - type: object
              properties:
                timestamp:
                  type: number
                  description: Heartbeat timestamp in milliseconds since epoch
              description: Heartbeat event data
            - type: object
              properties:
                reason:
                  type: string
                  enum: [user, timeout]
                  description: The reason for cancellation
              description: Cancelled event data

    McpServerConfig:
      type: object
      properties:
        command:
          type: string
          description: Command to start the server (CLI mode)
        args:
          type: array
          items:
            type: string
          description: Command line arguments (CLI mode)
        url:
          type: string
          description: Server URL (SSE mode)
        env:
          type: object
          additionalProperties:
            type: string
          description: Environment variables

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 400
              type:
                type: string
                example: "invalid_request"
              message:
                type: string
                example: "Validation error"
              details:
                type: array
                items:
                  type: object
                  properties:
                    code:
                      type: string
                    path:
                      type: array
                      items:
                        type: string
                    message:
                      type: string

paths:
  /agent/messages:
    get:
      summary: Get agent messages
      description: Returns all messages in the agent's history
      operationId: getAgentMessages
      responses:
        "200":
          description: Agent messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    delete:
      summary: Clear agent messages
      description: Clears all messages in the agent's history
      operationId: clearAgentMessages
      responses:
        "204":
          description: Messages cleared

  /agent/task/sse:
    post:
      summary: Run a task with Server-Sent Events
      description: |
        Runs a task using the agent. This endpoint uses server-sent events (SSE) to stream
        the agent's responses back to the client in real-time. The endpoint adds heartbeat events
        every 30 seconds during periods of inactivity to help keep the connection alive and detect disconnections.
      operationId: runTaskSSE
      requestBody:
        description: Task request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
      responses:
        "200":
          description: Task started
          content:
            text/event-stream:
              schema:
                type: string
                description: Stream of server-sent events containing task progress
              example: |
                event: content_delta
                data: {"eventId":"task_1650000000000_0","content":"I'll help you create a React component."}

                event: tool_use_delta
                data: {"eventId":"task_1650000000000_1","name":"read_file","partialInput":{"target_file":"src/components/Example.tsx"}}

                event: message
                data: {"eventId":"task_1650000000000_2","role":"assistant","content":[{"type":"text","text":"I'll help you create a React component."}]}

                event: heartbeat
                data: {"eventId":"task_1650000000000_3","timestamp":1650000030000}

                event: complete
                data: {"eventId":"task_1650000000000_4"}
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Another task is already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: Reconnect to a running task stream
      description: |
        Reconnects to an existing task stream. This endpoint allows clients to recover from connection
        interruptions by requesting all events that occurred after their last received event ID.
        The server will replay all events from the history that occurred after the specified event ID.
        If no task is running, returns 204 No Content.
      operationId: reconnectTaskSSE
      parameters:
        - name: lastEventId
          in: query
          schema:
            type: string
          description: The ID of the last event received by the client (format task_<timestamp>_<sequence>)
      responses:
        "200":
          description: Connected to running task stream
          content:
            text/event-stream:
              schema:
                type: string
                description: Stream of server-sent events containing task progress
        "204":
          description: No task is currently running

  /agent/task/cancel:
    post:
      summary: Cancel the current task
      description: Cancels the currently running task, if any
      operationId: cancelTask
      responses:
        "204":
          description: Task cancelled successfully
        "404":
          description: No task is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /agent/task/ws:
    get:
      summary: Run a task with WebSocket
      description: |
        Establishes a WebSocket connection for running tasks using the agent.
        The client sends task requests as JSON messages and receives stream events
        as responses.
      operationId: runTaskWebSocket
      responses:
        "101":
          description: WebSocket connection established

  /agent/checkpoints:
    get:
      summary: List checkpoints
      description: Returns all available checkpoints for the current workspace
      operationId: listCheckpoints
      responses:
        "200":
          description: List of checkpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Checkpoint"

  /agent/checkpoints/{checkpointId}/apply:
    parameters:
      - name: checkpointId
        in: path
        required: true
        schema:
          type: string
          minLength: 1
        description: ID of the checkpoint (Git commit hash)
    post:
      summary: Apply checkpoint
      description: Applies the changes from the specified checkpoint and updates message history
      operationId: applyCheckpoint
      responses:
        "200":
          description: Checkpoint applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  id:
                    type: string
                required:
                  - success
                  - id
        "400":
          $ref: "#/components/responses/ValidationError"

  /health:
    get:
      summary: Health check
      description: Returns the health status of the API
      operationId: healthCheck
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok]
                  version:
                    type: string
                  uptime:
                    type: number
                    format: float

  /mcp/servers:
    get:
      summary: List all registered MCP servers
      description: Returns all registered MCP servers with their tools
      operationId: listMcpServers
      responses:
        "200":
          description: List of MCP servers
          content:
            application/json:
              schema:
                type: object
                properties:
                  servers:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: Unique identifier for the server
                        enabled:
                          type: boolean
                          description: Whether the server is enabled
                        tools:
                          type: array
                          items:
                            type: string
                          description: List of tool names provided by this server

  /mcp/register:
    post:
      summary: Register new MCP servers
      description: Registers one or more MCP servers with their configurations
      operationId: registerMcpServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                $ref: "#/components/schemas/McpServerConfig"
      responses:
        "201":
          description: Servers registered successfully
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Server already exists
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - type
                  - message
                properties:
                  code:
                    type: integer
                    example: 409
                  type:
                    type: string
                    example: "already_exists"
                  message:
                    type: string
                    description: Error message explaining which server already exists
                  details:
                    type: object
                    description: Additional error details

  /mcp/servers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
        description: ID of the MCP server
    get:
      summary: Get MCP server configuration
      description: Returns the configuration of a specific MCP server
      operationId: getMcpServerConfig
      responses:
        "200":
          description: Server configuration
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/McpServerConfig"
        "400":
          $ref: "#/components/responses/ValidationError"
    delete:
      summary: Deregister a MCP server
      description: Removes a registered MCP server
      operationId: deregisterMcpServer
      responses:
        "204":
          description: Server deregistered successfully
        "400":
          $ref: "#/components/responses/ValidationError"
    put:
      summary: Update MCP server configuration
      description: Updates the configuration of an existing MCP server
      operationId: updateMcpServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                $ref: "#/components/schemas/McpServerConfig"
      responses:
        "204":
          description: Server configuration updated successfully
        "400":
          $ref: "#/components/responses/ValidationError"

  /mcp/servers/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
        description: ID of the MCP server
    put:
      summary: Update server status
      description: Enable or disable a specific MCP server
      operationId: updateMcpServerStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - enabled
              properties:
                enabled:
                  type: boolean
                  description: Whether to enable or disable the server
      responses:
        "204":
          description: Server status updated successfully
        "400":
          $ref: "#/components/responses/ValidationError"

  /mcp/tools:
    get:
      summary: List available tools
      description: Returns all tools available from registered MCP servers
      operationId: listMcpTools
      responses:
        "200":
          description: List of available tools
          content:
            application/json:
              schema:
                type: object
                properties:
                  tools:
                    type: array
                    items:
                      type: string
                      description: Name of the tool
                required:
                  - tools

  /mcp/reload:
    get:
      summary: Reload MCP configuration
      description: Reloads the MCP configuration and reinitializes all servers
      operationId: reloadMcpConfig
      responses:
        "200":
          description: Configuration reloaded successfully